requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in setup.py
#
appdirs==1.4.3            # via black
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, flake8-bugbear, pytest
black==19.3b0
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via black
coverage==4.5.3           # via pytest-cov
entrypoints==0.3          # via flake8
flake8-bugbear==19.3.0
flake8-comprehensions==2.1.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2    # via flake8-docstrings
flake8-quotes==2.0.1
flake8==3.7.7
idna==2.8                 # via requests
importlib-metadata==0.18  # via pluggy, pytest
mccabe==0.6.1             # via flake8
more-itertools==7.0.0     # via pytest
packaging==19.0           # via pytest
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pydantic==0.23
pydocstyle==3.0.0         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
pyhumps==1.2.2
pyparsing==2.4.0          # via packaging
pytest-cov==2.7.1
pytest==4.6.3
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0               # via packaging, pydocstyle, pytest, python-dateutil
snowballstemmer==1.2.1    # via pydocstyle
toml==0.10.0              # via black
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via pytest
zipp==0.5.1               # via importlib-metadata
